refactor(api): extract webhooks subdomain from system.rs (#3749)
