+++ crates/librefang-api/src/routes/mod.rs
@@ -33,6 +33,7 @@ pub mod skills;
 pub mod system;
 pub mod terminal;
 pub mod users;
+pub mod webhooks;
 pub mod workflows;
 
 // Glob re-export to keep `routes::handler_name` backward compatible
@@ -64,6 +65,7 @@ pub use skills::*;
 pub use system::*;
 pub use terminal::*;
 pub use users::*;
+pub use webhooks::*;
 pub use workflows::*;
 
 use crate::middleware::RequestLanguage;
